Revert "Rename djangoapp app (#575)"
This reverts commit 74b2b61c04817af1f4706b88be0a3267660a6b05.
