+++ queue/management/commands/count_queued_submissions.py
@@ -7,7 +7,7 @@ from __future__ import unicode_literals
 from __future__ import absolute_import
 from __future__ import print_function
 
-from submission_queue.models import Submission
+from queue.models import Submission
 
 import backoff
 import boto3
