+++ queue/management/commands/retire_failed_submissions.py
@@ -1,8 +1,8 @@
 from __future__ import absolute_import
 import logging
 from optparse import make_option
-from submission_queue.consumer import post_failure_to_lms
-from submission_queue.models import Submission
+from queue.consumer import post_failure_to_lms
+from queue.models import Submission
 
 from django.conf import settings
 from django.core.management.base import BaseCommand
