+++ queue/management/commands/tests/test_count_queued_submissions.py
@@ -1,5 +1,5 @@
 from __future__ import absolute_import
-from submission_queue.models import Submission
+from queue.models import Submission
 
 from django.core.management import call_command
 from django.test import TestCase
