+++ queue/management/commands/tests/test_retire_failed_submissions.py
@@ -13,7 +13,7 @@ class TestRetireFailedSubmissions(TestCase):
     Tests of the retire_submissions management command.
     """
     def test_all_queues(self):
-        path = 'submission_queue.management.commands.retire_failed_submissions.Command.retire_submissions'
+        path = 'queue.management.commands.retire_failed_submissions.Command.retire_submissions'
         with mock.patch(path) as mock_method:
             call_command(u'retire_failed_submissions', force=True)
             assert mock_method.call_count == 1
