+++ queue/management/commands/tests/test_run_consumer.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import
-from submission_queue.consumer import Worker
-from submission_queue.management.commands.run_consumer import Command
+from queue.consumer import Worker
+from queue.management.commands.run_consumer import Command
 
 from django.core.management import call_command
 from django.test import TestCase, override_settings
@@ -18,7 +18,7 @@ class RunConsumer(TestCase):
         call_command('run_consumer')
 
     @patch.object(Worker, 'run')
-    @patch('submission_queue.management.commands.run_consumer.MONITOR_SLEEPTIME', 0)
+    @patch('queue.management.commands.run_consumer.MONITOR_SLEEPTIME', 0)
     def test_stop_workers(self, mock_worker):
         """
         Having a Worker() with a zero exitcode tells monitor
