+++ queue/urls.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import
-from submission_queue.ext_interface import get_queuelen, get_submission, put_result
-from submission_queue.lms_interface import submit
-from submission_queue.views import log_in, log_out, status
+from queue.ext_interface import get_queuelen, get_submission, put_result
+from queue.lms_interface import submit
+from queue.views import log_in, log_out, status
 
 from django.conf.urls import url
