+++ xqueue/settings.py
@@ -116,7 +116,7 @@ INSTALLED_APPS = (
     # 'django.contrib.admin',
     # Uncomment the next line to enable admin documentation:
     # 'django.contrib.admindocs',
-    'submission_queue',
+    'queue',
     'release_util',
     'storages',
 )
