+++ xqueue/urls.py
@@ -2,5 +2,5 @@ from __future__ import absolute_import
 from django.conf.urls import include, url
 
 urlpatterns = [
-    url(r'^xqueue/', include('submission_queue.urls')),
+    url(r'^xqueue/', include('queue.urls')),
 ]
